+++ passes/V009/testdata/src/a/outside_package.go
@@ -1,0 +1,10 @@
+package a
+
+import (
+	"a/validation"
+	"regexp"
+)
+
+func foutside() {
+	validation.StringMatch(regexp.MustCompile(`^[a-zA-Z0-9.-]+$`), "")
+}